[Refactor] Change TourItem to sealed interface
